Create Topic Models from 1 to k
Source: R/topic_modelling.R

Returns a list, where every position is a created model. List of k values must be provided as a vector.

CalculateLDAModelsInKSet(folder, Ks = 2:20, months, method = "VEM")

Arguments

folder
Folder variable containing all raw data. See loadFiles.

Ks
A vector of integers (e.g. c(1,3,2,7)) specifying the number of topics to be created

months
A character vector containing the months of interest (currently one uses the first month)

Value

A list of the models for all chosen values in Ks